Allow clients to register custom encodings
A client may wish to specify a custom pseudo-encoding to signal some
kind of capability or feature support to the server.
It may also wish to register a handler for a custom encoding (most
likely not for image decoding, but for sending other non-image messages)

To facilitate this, add a customEncodings attribute to the RFB options
object. If specified, it contains a list of encoding ID's, and
optionally a handler to be called when receiving messages with this ID.
If no handler is specified, the encoding is a pseudo-encoding.